Removes Proximity Normals
