+++ src/fighter/ryu/status/dash_back.rs
@@ -1,0 +1,21 @@
+use {
+    crate::imports::status_imports::*,
+    crate::fighter::common::status::movement::dash::*
+};
+
+#[status_script(agent = "ryu", status = FIGHTER_RYU_STATUS_KIND_DASH_BACK, condition = LUA_SCRIPT_STATUS_FUNC_STATUS_PRE)]
+unsafe fn ryu_dashback_pre(fighter: &mut L2CFighterCommon) -> L2CValue {
+    fgc_dashback_pre(fighter)
+}
+
+#[status_script(agent = "ryu", status = FIGHTER_RYU_STATUS_KIND_DASH_BACK, condition = LUA_SCRIPT_STATUS_FUNC_STATUS_MAIN)]
+unsafe fn ryu_dashback_main(fighter: &mut L2CFighterCommon) -> L2CValue {
+    fgc_dashback_main(fighter)
+}
+
+pub fn install() {
+    install_status_scripts!(
+        ryu_dashback_pre,
+        ryu_dashback_main
+    );
+}
